Keeping up with the upstream (#1)
* Logger enhancements, see #1882 (#1889)

* Add ability to pass in mdns configuration

* Callback before homebridge exits (#1907)

* Logger enhancements, see #1882

* Callback before homebridge exits, see #1905

* Remove package-lock.json

* 0.4.44

* prevent npm update check warnings (#1849)

* Option to hide QRcode in logs (#1911)

* Option to hide QRcode in logs

* add no-qrcode parameter

* declare showQRCode

* consolidate insecureAccess as a opts

* consolidate insecureAccess into opts

* More correct sentences

* invert hideQRCode showQRCode logic (#1913)

* invert hideQRCode showQRCode logic

* invert _hideQRCode logic

* Set theme jekyll-theme-cayman

* Create CNAME

* Handle both CommonJS and ES6 style module.exports.

In ES6/Babel 6 doing `export default function() {}` results in that
function being assigned to `module.exports.default` rather than to
`module.exports` directly as it does in CommonJS.

Currently the Homebridge platform plugin setup expects the initializer
function to be available from `require` in CommonJS style, and so it
breaks for projects using ES6. Additionally there is currently no
helpful error when this happens; Homebridge fails to boot with Node
reporting a JS type error that `plugin.initializer` is not a function.

Here I've retained backwards compatibility with CommonJS style, but
check to see if `module.exports` is a function itself or if we need to
access `module.exports.default`. If neither is a function, we throw a
helpful error.

* Version bump for npm

* Use logo from Slack

* Use 2x logo

* Schmancy website

Modified from @taitems's beautiful site and designs.

* Hush Github's deps warnings

* Logo "draws itself"

* Add footer

* Almost forgot design+logo attribution!!

* Updates to README on developing plugins (#2052)

I wanted to add a quick summary about the differences between platform
and single acccessories, and then also grouped the information
accordingly.

* Add link to ES6 sample plugin

* Fix slack inviter.

Closes #2049

* Update ISSUE_TEMPLATE.md (#2107)

* Update ISSUE_TEMPLATE.md

* Update ISSUE_TEMPLATE.md

* Update hap-nodejs to support TV profile, and update serialization process to persist linked services.

* v0.4.46
TV Support

* Add Discord community links (#2129)

* Update README.md

* Update README.md

* Minor README update

* Resolve custom config path so plugins extracting it can work with it

* Remove clashing vars names

* Add ability to publish external accessory separately.

* 0.4.47

* Add ports to config sample

* Added an item to the list of common issues (#2186)

This is based on: https://alwayswithyouwhatcannotbedone.wordpress.com/2019/03/10/temporary-fix-for-an-iphone-that-cannot-access-local-addresses/

* Clean orphaned cached accessories / handle uncaught exceptions (#2190)

* Update cli.js

Add `-R` or `--remove-orphans` command line options to clean orphaned cached accessories for which no plugin is found.  See #2120.

* Update server.js

Clean orphaned cached accessores, for which no plugin was loaded, when `-R` was specified on the command line.  See #2120.

* Update cli.js

Handle uncaught exceptions to do a clean shutdown instead of crashing.

* Update dependency version

* 0.4.48

* Node.js 12 support (#2205)

* Update dependencies

* Switch back node dependencies

* 0.4.49

* Change hosting of Logo to homebridge itself (#2229)

Current one was raising a security issue with Chrome

Placed a copy of the image in #1865

* 0.4.50